Creamos auxiliar.php, creamos la Accion: Borrar, con comprobaciones de que exista la pelicula

+++ index.php
@@ -6,14 +6,23 @@
     </head>
     <body>
         <?php
-        $pdo = new PDO('pgsql:host=localhost;dbname=fa','fa','fa');
+        require 'auxiliar.php';
+        $pdo = conectar();
         //Pregunto si vengo del confirm_borrado, si existe un id por POST, es que quiero borrar una fila
         if (isset($_POST['id'])) {
             $id = $_POST['id'];
-            $st = $pdo->prepare('DELETE FROM peliculas WHERE id = :id');
-            $st->execute([':id' => $id]); ?>
-            <h3>Película borrada correctamente.</h3>
+            $pdo->beginTransaction();
+            $pdo->exec('LOCK TABLE peliculas IN SHARE MODE');
+            if (!buscarPelicula($pdo, $id)) { ?>
+                <h3>Error: La pelicula no existe!</h3>
+            <?php
+            } else {
+                $st = $pdo->prepare('DELETE FROM peliculas WHERE id = :id');
+                $st->execute([':id' => $id]); ?>
+                <h3>Película borrada correctamente.</h3>
         <?php
+            }
+            $pdo->commit();
         }
         $buscarTitulo = isset($_GET['buscarTitulo'])
                         ? trim($_GET['buscarTitulo'])
